Crate sui_config

Re-exports

pub use node::ConsensusConfig;

pub use node::NodeConfig;

pub use node::ValidatorInfo;

Modules

builder

genesis

genesis_config

node

p2p

utils

Structs

FullnodeConfigBuilder

NetworkConfig

This is a config that is used for testing or local use as it contains the config and keys for all validators

PersistedConfig

Constants

AUTHORITIES_DB_NAME

CONSENSUS_DB_NAME

FULL_NODE_DB_PATH

SUI_CLIENT_CONFIG

SUI_DEV_NET_URL

SUI_FULLNODE_CONFIG

SUI_GENESIS_FILENAME

SUI_KEYSTORE_FILENAME

SUI_NETWORK_CONFIG

Traits

Config

Functions

sui_config_dir