sui_config::node
Struct NodeConfig

pub struct NodeConfig {

    pub protocol_key_pair: AuthorityKeyPairWithPath,
    pub worker_key_pair: KeyPairWithPath,
    pub account_key_pair: KeyPairWithPath,
    pub network_key_pair: KeyPairWithPath,
    pub db_path: PathBuf,
    pub network_address: Multiaddr,
    pub json_rpc_address: SocketAddr,
    pub rpc: Option<Config>,
    pub metrics_address: SocketAddr,
    pub admin_interface_port: u16,
    pub consensus_config: Option<ConsensusConfig>,
    pub enable_index_processing: bool,
    pub remove_deprecated_tables: bool,
    pub jsonrpc_server_type: Option<ServerType>,
    pub grpc_load_shed: Option<bool>,
    pub grpc_concurrency_limit: Option<usize>,
    pub p2p_config: P2pConfig,
    pub genesis: Genesis,
    pub authority_store_pruning_config: AuthorityStorePruningConfig,
    pub end_of_epoch_broadcast_channel_capacity: usize,
    pub checkpoint_executor_config: CheckpointExecutorConfig,
    pub metrics: Option<MetricsConfig>,
    pub supported_protocol_versions: Option<SupportedProtocolVersions>,
    pub db_checkpoint_config: DBCheckpointConfig,
    pub expensive_safety_check_config: ExpensiveSafetyCheckConfig,
    pub name_service_package_address: Option<SuiAddress>,
    pub name_service_registry_id: Option<ObjectID>,
    pub name_service_reverse_registry_id: Option<ObjectID>,
    pub transaction_deny_config: TransactionDenyConfig,
    pub certificate_deny_config: CertificateDenyConfig,
    pub state_debug_dump_config: StateDebugDumpConfig,
    pub state_archive_read_config: Vec<StateArchiveConfig>,
    pub state_snapshot_write_config: StateSnapshotConfig,
    pub indexer_max_subscriptions: Option<usize>,
    pub transaction_kv_store_read_config: TransactionKeyValueStoreReadConfig,
    pub transaction_kv_store_write_config: Option<TransactionKeyValueStoreWriteConfig>,
    pub jwk_fetch_interval_seconds: u64,
    pub zklogin_oauth_providers: BTreeMap<Chain, BTreeSet<String>>,
    pub authority_overload_config: AuthorityOverloadConfig,
    pub run_with_range: Option<RunWithRange>,
    pub policy_config: Option<PolicyConfig>,
    pub firewall_config: Option<RemoteFirewallConfig>,
    pub execution_cache: ExecutionCacheConfig,
    pub state_accumulator_v2: bool,
    pub enable_soft_bundle: bool,
    pub enable_validator_tx_finalizer: bool,
    pub verifier_signing_config: VerifierSigningConfig,
    pub enable_db_write_stall: Option<bool>,
    pub execution_time_observer_config: Option<ExecutionTimeObserverConfig>,
    pub chain_override_for_testing: Option<Chain>,

}

Fields
protocol_key_pair: AuthorityKeyPairWithPathworker_key_pair: KeyPairWithPathaccount_key_pair: KeyPairWithPathnetwork_key_pair: KeyPairWithPathdb_path: PathBufnetwork_address: Multiaddrjson_rpc_address: SocketAddrrpc: Option<Config>metrics_address: SocketAddradmin_interface_port: u16consensus_config: Option<ConsensusConfig>enable_index_processing: boolremove_deprecated_tables: booljsonrpc_server_type: Option<ServerType>

Determines the jsonrpc server type as either:

‘websocket’ for a websocket based service (deprecated)

‘http’ for an http based service

‘both’ for both a websocket and http based service (deprecated)

grpc_load_shed: Option<bool>grpc_concurrency_limit: Option<usize>p2p_config: P2pConfiggenesis: Genesisauthority_store_pruning_config: AuthorityStorePruningConfigend_of_epoch_broadcast_channel_capacity: usize

Size of the broadcast channel used for notifying other systems of end of epoch.

If unspecified, this will default to 128.

checkpoint_executor_config: CheckpointExecutorConfigmetrics: Option<MetricsConfig>supported_protocol_versions: Option<SupportedProtocolVersions>

In a sui-node binary, this is set to SupportedProtocolVersions::SYSTEM_DEFAULT in sui-node/src/main.rs. It is present in the config so that it can be changed by tests in order to test protocol upgrades.

db_checkpoint_config: DBCheckpointConfigexpensive_safety_check_config: ExpensiveSafetyCheckConfigname_service_package_address: Option<SuiAddress>name_service_registry_id: Option<ObjectID>name_service_reverse_registry_id: Option<ObjectID>transaction_deny_config: TransactionDenyConfigcertificate_deny_config: CertificateDenyConfigstate_debug_dump_config: StateDebugDumpConfigstate_archive_read_config: Vec<StateArchiveConfig>state_snapshot_write_config: StateSnapshotConfigindexer_max_subscriptions: Option<usize>transaction_kv_store_read_config: TransactionKeyValueStoreReadConfigtransaction_kv_store_write_config: Option<TransactionKeyValueStoreWriteConfig>jwk_fetch_interval_seconds: u64zklogin_oauth_providers: BTreeMap<Chain, BTreeSet<String>>authority_overload_config: AuthorityOverloadConfigrun_with_range: Option<RunWithRange>policy_config: Option<PolicyConfig>firewall_config: Option<RemoteFirewallConfig>execution_cache: ExecutionCacheConfigstate_accumulator_v2: boolenable_soft_bundle: boolenable_validator_tx_finalizer: boolverifier_signing_config: VerifierSigningConfigenable_db_write_stall: Option<bool>

If a value is set, it determines if writes to DB can stall, which can halt the whole process. By default, write stall is enabled on validators but not on fullnodes.

execution_time_observer_config: Option<ExecutionTimeObserverConfig>chain_override_for_testing: Option<Chain>

Allow overriding the chain for testing purposes. For instance, it allows you to create a test network that believes it is mainnet or testnet. Attempting to override this value on production networks will result in an error.

Implementations

impl NodeConfig

pub fn protocol_key_pair(&self) -> &AuthorityKeyPair

pub fn worker_key_pair(&self) -> &NetworkKeyPair

pub fn network_key_pair(&self) -> &NetworkKeyPair

pub fn protocol_public_key(&self) -> AuthorityPublicKeyBytes

pub fn db_path(&self) -> PathBuf

pub fn db_checkpoint_path(&self) -> PathBuf

pub fn archive_path(&self) -> PathBuf

pub fn snapshot_path(&self) -> PathBuf

pub fn network_address(&self) -> &Multiaddr

pub fn consensus_config(&self) -> Option<&ConsensusConfig>

pub fn genesis(&self) -> Result<&Genesis>

pub fn sui_address(&self) -> SuiAddress

pub fn archive_reader_config(&self) -> Option<ArchiveReaderConfig>

pub fn jsonrpc_server_type(&self) -> ServerType

pub fn rpc(&self) -> Option<&Config>