sui_config

Trait Config

pub trait Config
where
    Self: DeserializeOwned + Serialize,
{
    // Provided methods
    fn persisted(self, path: &Path) -> PersistedConfig<Self> { ... }
    fn load<P: AsRef<Path>>(path: P) -> Result<Self, Error> { ... }
    fn save<P: AsRef<Path>>(&self, path: P) -> Result<(), Error> { ... }
    fn load_with_lock<P: AsRef<Path>>(path: P) -> Result<Self, Error> { ... }
    fn save_with_lock<P: AsRef<Path>>(&self, path: P) -> Result<(), Error> { ... }
}

Provided Methods

fn persisted(self, path: &Path) -> PersistedConfig<Self>

fn load<P: AsRef<Path>>(path: P) -> Result<Self, Error>

fn save<P: AsRef<Path>>(&self, path: P) -> Result<(), Error>

fn load_with_lock<P: AsRef<Path>>(path: P) -> Result<Self, Error>

Load the config from the given path, acquiring a shared lock on the file during the read.

fn save_with_lock<P: AsRef<Path>>(&self, path: P) -> Result<(), Error>

Save the config to the given path, acquiring an exclusive lock on the file during the write.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.