Module sui_config::utils

Functions

available_local_socket_address

available_network_socket_address

get_available_port

Return an ephemeral, available port. On unix systems, the port returned will be in the TIME_WAIT state ensuring that the OS won’t hand out this port for some grace period. Callers should be able to bind to this port given they use SO_REUSEADDR.

get_local_ip_for_tests

new_tcp_network_address

new_udp_network_address

socket_address_to_udp_multiaddr

udp_multiaddr_to_listen_address