sui_config
Module verifier_signing_config

Structs

VerifierSigningConfig

This holds limits that are only set and used by the verifier during signing only. There are additional limits in the MeterConfig and VerifierConfig that are used during both signing and execution, however those limits cannot be set here and must be protocol versioned.

Constants

DEFAULT_MAX_BACK_EDGES_PER_FUNCTION

DEFAULT_MAX_BACK_EDGES_PER_MODULE

DEFAULT_MAX_PER_FUN_METER_UNITS

DEFAULT_MAX_PER_MOD_METER_UNITS

DEFAULT_MAX_PER_PKG_METER_UNITS