Struct sui_config::node::NodeConfig

pub struct NodeConfig {

    pub protocol_key_pair: AuthorityKeyPairWithPath,
    pub worker_key_pair: KeyPairWithPath,
    pub account_key_pair: KeyPairWithPath,
    pub network_key_pair: KeyPairWithPath,
    pub db_path: PathBuf,
    pub network_address: Multiaddr,
    pub json_rpc_address: SocketAddr,
    pub enable_experimental_rest_api: bool,
    pub metrics_address: SocketAddr,
    pub admin_interface_port: u16,
    pub consensus_config: Option<ConsensusConfig>,
    pub enable_event_processing: bool,
    pub enable_index_processing: bool,
    pub grpc_load_shed: Option<bool>,
    pub grpc_concurrency_limit: Option<usize>,
    pub p2p_config: P2pConfig,
    pub genesis: Genesis,
    pub authority_store_pruning_config: AuthorityStorePruningConfig,
    pub end_of_epoch_broadcast_channel_capacity: usize,
    pub checkpoint_executor_config: CheckpointExecutorConfig,
    pub metrics: Option<MetricsConfig>,
    pub supported_protocol_versions: Option<SupportedProtocolVersions>,
    pub db_checkpoint_config: DBCheckpointConfig,
    pub indirect_objects_threshold: usize,
    pub expensive_safety_check_config: ExpensiveSafetyCheckConfig,
    pub name_service_package_address: Option<SuiAddress>,
    pub name_service_registry_id: Option<ObjectID>,
    pub name_service_reverse_registry_id: Option<ObjectID>,
    pub transaction_deny_config: TransactionDenyConfig,
    pub certificate_deny_config: CertificateDenyConfig,
    pub state_debug_dump_config: StateDebugDumpConfig,
    pub state_archive_write_config: StateArchiveConfig,
    pub state_archive_read_config: Vec<StateArchiveConfig>,
    pub state_snapshot_write_config: StateSnapshotConfig,
    pub indexer_max_subscriptions: Option<usize>,
    pub transaction_kv_store_read_config: TransactionKeyValueStoreReadConfig,
    pub transaction_kv_store_write_config: Option<TransactionKeyValueStoreWriteConfig>,
    pub jwk_fetch_interval_seconds: u64,
    pub zklogin_oauth_providers: BTreeMap<Chain, BTreeSet<String>>,
    pub overload_threshold_config: OverloadThresholdConfig,

}

Fields
protocol_key_pair: AuthorityKeyPairWithPathworker_key_pair: KeyPairWithPathaccount_key_pair: KeyPairWithPathnetwork_key_pair: KeyPairWithPathdb_path: PathBufnetwork_address: Multiaddrjson_rpc_address: SocketAddrenable_experimental_rest_api: boolmetrics_address: SocketAddradmin_interface_port: u16consensus_config: Option<ConsensusConfig>enable_event_processing: boolenable_index_processing: boolgrpc_load_shed: Option<bool>grpc_concurrency_limit: Option<usize>p2p_config: P2pConfiggenesis: Genesisauthority_store_pruning_config: AuthorityStorePruningConfigend_of_epoch_broadcast_channel_capacity: usize

Size of the broadcast channel used for notifying other systems of end of epoch.

If unspecified, this will default to 128.

checkpoint_executor_config: CheckpointExecutorConfigmetrics: Option<MetricsConfig>supported_protocol_versions: Option<SupportedProtocolVersions>

In a sui-node binary, this is set to SupportedProtocolVersions::SYSTEM_DEFAULT in sui-node/src/main.rs. It is present in the config so that it can be changed by tests in order to test protocol upgrades.

db_checkpoint_config: DBCheckpointConfigindirect_objects_threshold: usizeexpensive_safety_check_config: ExpensiveSafetyCheckConfigname_service_package_address: Option<SuiAddress>name_service_registry_id: Option<ObjectID>name_service_reverse_registry_id: Option<ObjectID>transaction_deny_config: TransactionDenyConfigcertificate_deny_config: CertificateDenyConfigstate_debug_dump_config: StateDebugDumpConfigstate_archive_write_config: StateArchiveConfigstate_archive_read_config: Vec<StateArchiveConfig>state_snapshot_write_config: StateSnapshotConfigindexer_max_subscriptions: Option<usize>transaction_kv_store_read_config: TransactionKeyValueStoreReadConfigtransaction_kv_store_write_config: Option<TransactionKeyValueStoreWriteConfig>jwk_fetch_interval_seconds: u64zklogin_oauth_providers: BTreeMap<Chain, BTreeSet<String>>overload_threshold_config: OverloadThresholdConfig
Implementations

impl NodeConfig

pub fn protocol_key_pair(&self) -> &AuthorityKeyPair

pub fn worker_key_pair(&self) -> &NetworkKeyPair

pub fn network_key_pair(&self) -> &NetworkKeyPair

pub fn protocol_public_key(&self) -> AuthorityPublicKeyBytes

pub fn db_path(&self) -> PathBuf

pub fn db_checkpoint_path(&self) -> PathBuf

pub fn archive_path(&self) -> PathBuf

pub fn snapshot_path(&self) -> PathBuf

pub fn network_address(&self) -> &Multiaddr

pub fn consensus_config(&self) -> Option<&ConsensusConfig>

pub fn genesis(&self) -> Result<&Genesis>

pub fn sui_address(&self) -> SuiAddress

pub fn archive_reader_config(&self) -> Vec<ArchiveReaderConfig>